set up the framework for a GUI to plot parameters selected from a pick box.
The GUI will allow loading two files, one a data file with time series signals,
the other a log file from Calterm III. The pick box will be populated from
the file header with the names of the parameters. The pick box will allow
selection of the parameters to display in a plot.