applied individual channel gains to the plotted value for each channel changed the placement of the parameter name from title to ylabel
Note: change to make: make the parameter and channel data belong to
the parameter and channel classes. Then store the list of selected
channels in a way that allows selection of the class member. As it is,
the data lives as a separate entity and the class only holds the
metadata. Thus, separate properties are required for the selected
parameters, the selected channels, the selected gains, etc. Selecting
the class instances themselves would be a much more elegant solution.